[EDIFIKANA] Enabling previews in the JVM target
