+++ edifikana/front-end/shared-compose/src/jvmMain/kotlin/com/cramsan/edifikana/client/lib/features/root/auth/signinv2/SignInV2ContentPreview.kt
@@ -1,4 +1,4 @@
-package com.cramsan.edifikana.client.lib.features.signinv2
+package com.cramsan.edifikana.client.lib.features.root.auth.signinv2
 
 import androidx.compose.desktop.ui.tooling.preview.Preview
 import androidx.compose.runtime.Composable
