+++ edifikana/front-end/shared-compose/src/jvmMain/kotlin/com/cramsan/edifikana/client/lib/features/root/main/timecard/stafflist/StaffListScreenPreview.kt
@@ -1,0 +1,25 @@
+package com.cramsan.edifikana.client.lib.features.root.main.timecard.stafflist
+
+import androidx.compose.desktop.ui.tooling.preview.Preview
+import androidx.compose.runtime.Composable
+import com.cramsan.edifikana.lib.model.StaffId
+
+@Preview
+@Composable
+private fun StaffListScreenPreview() {
+    StaffList(
+        isLoading = true,
+        staffs = listOf(
+            StaffUIModel(
+                "Cesar Andres Ramirez Sanchez",
+                StaffId("John"),
+            ),
+            StaffUIModel(
+                "2",
+                StaffId("Jane"),
+            ),
+        ),
+        onStaffClick = {},
+        onAddStaffClick = {},
+    )
+}
