+++ edifikana/front-end/shared-compose/src/jvmMain/kotlin/com/cramsan/edifikana/client/lib/features/root/splash/SplashActivityScreenPreview.kt
@@ -1,0 +1,13 @@
+package com.cramsan.edifikana.client.lib.features.root.splash
+
+import androidx.compose.desktop.ui.tooling.preview.Preview
+import androidx.compose.runtime.Composable
+
+/**
+ * Preview for the SplashActivity screen.
+ */
+@Preview
+@Composable
+private fun SplashActivityScreenPreview() {
+    SplashActivityScreen()
+}
